Revert "Add send emails feature , currenty logged in user" (#48)
This reverts commit 7ec31dcb79714dc1ac77f1aa90611f3ac5545ecb.

+++ src/main/java/com/featuredoc/models/Feature.java
@@ -14,8 +14,8 @@ public class Feature {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long featureID;
-    private @NonNull Long createdBy;
+    private Integer featureID;
+    private @NonNull Integer createdBy;
     @Column(updatable = false, insertable = false)
     private Timestamp createdAt;
 }
